+++ changelog.md
@@ -423,3 +423,6 @@
 ### D15
 - animation glitch bug decreased
 - D15: disabled debugging mode
+
+### D16
+- started page ui and animations updated